Fix check lines
To avoid reporting getSite by getSiteManager some logic was added on
``check_line`` the check for that.

But since some time ago an opening parethensis was added on most of the replacments
that logic was working against us.

The proper fix would be to use an AST parser.

This patch is a quick fix before the proper one.